Added L3 to linker command file.
Added L3 usage.
Added EMMC code.

+++ AM335X_StarterWare_02_00_01_01/include/ramdisk.h
@@ -42,5 +42,5 @@
 void RAM_disk_read(unsigned int lba, unsigned char *buf, unsigned int len);
 void RAM_disk_write(unsigned int lba, unsigned char *buf, unsigned int len);
 void RAM_disk_initialize(void);
-void RAM_disk_ioctl (unsigned int drive, unsigned int  command,  unsigned int* buffer);
+void RAM_disk_ioctl (unsigned int drive, unsigned int  command,  unsigned long* buffer);
